Fixed a typo in the  and Added a  file for instructions to connect the bluetooth module
Changed the names in `STD_TYBES` to ba as like as stdint to work with
avr-gcc compiler.

Changed the names in `STD_TYBES` to ba as like as stdint to work with
avr-gcc compiler.

Changed the names in  to ba as like as stdint.

Changed the STD_TYPES to be compatible with avr-gcc compiler

Changed the names in  to ba as like as stdint.

Changed the STD_TYPES to be compatible with avr-gcc compiler

Changed the names in `STD_TYBES` to ba as like as stdint to work with
avr-gcc compiler.

Changed the names in  to ba as like as stdint.

Changed the STD_TYPES to be compatible with avr-gcc compiler

Changed the names in `STD_TYBES` to ba as like as stdint to work with
avr-gcc compiler.

Changed the names in `STD_TYBES` to ba as like as stdint to work with
avr-gcc compiler.

Changed the names in  to ba as like as stdint.

Changed the STD_TYPES to be compatible with avr-gcc compiler

Changed the names in  to ba as like as stdint.

Changed the STD_TYPES to be compatible with avr-gcc compiler

Changed the names in `STD_TYBES` to ba as like as stdint to work with
avr-gcc compiler.

Changed the names in  to ba as like as stdint.

Changed the STD_TYPES to be compatible with avr-gcc compiler

Changed the names in `STD_TYBES` to ba as like as stdint to work with
avr-gcc compiler.

Fixed a typo in the  and Added a  file for instructions to connect the bluetooth module

Changed the names in `STD_TYBES` to ba as like as stdint to work with
avr-gcc compiler.